app/requirements.txt
Flask==2.2.5
openai==1.54.3
python-dotenv==0.21.0
kubernetes==31.0.0